requirements.txt
fastapi[standart]
uvicorn
python-dotenv
google-generativeai
opencv-python-headless
numpy
onnxruntime
pillow
torch
scipy
ultralytics
python-multipart
edge_tts